[cga] moved SDK-wide types to common/base
