+++ CMakeLists.txt
@@ -71,7 +71,6 @@ include(cmake/CondaEnv.cmake)
 
 # Add ClaraGenomicsAnalysis projects.
 add_subdirectory(common/config)
-add_subdirectory(common/defs)
 add_subdirectory(common/logging)
 add_subdirectory(common/base)
 add_subdirectory(common/io)
