+++ cudamapper/CMakeLists.txt
@@ -25,7 +25,7 @@ endif()
 cuda_add_library(minimizer
         src/minimizer.cu)
 target_include_directories(minimizer PUBLIC include ${CUB_DIR})
-target_link_libraries(minimizer cgadefs cgalogging pthread cgabase)
+target_link_libraries(minimizer cgabase cgalogging pthread)
 target_compile_options(minimizer PRIVATE -Werror)
 
 cuda_add_library(index_gpu
@@ -34,25 +34,25 @@ cuda_add_library(index_gpu
         src/index_host_copy.cu
         src/minimizer.cu)
 target_include_directories(index_gpu PUBLIC include ${CUB_DIR})
-target_link_libraries(index_gpu cgadefs cgalogging minimizer pthread cgabase cgaio)
+target_link_libraries(index_gpu cgabase cgaio cgalogging minimizer pthread)
 target_compile_options(index_gpu PRIVATE -Werror)
 
 cuda_add_library(matcher_gpu
         src/matcher_gpu.cu)
 target_include_directories(matcher_gpu PUBLIC include ${CUB_DIR})
-target_link_libraries(matcher_gpu cgadefs cgalogging cgabase cgaio)
+target_link_libraries(matcher_gpu cgabase cgaio cgalogging)
 target_compile_options(matcher_gpu PRIVATE -Werror)
 
 add_library(cudamapper_utils
         src/cudamapper_utils.cpp)
 target_include_directories(cudamapper_utils PUBLIC include)
-target_link_libraries(cudamapper_utils cgadefs)
+target_link_libraries(cudamapper_utils cgabase)
 
 cuda_add_library(overlapper_triggerred
         src/overlapper_triggered.cu
         src/overlapper.cpp)
 target_include_directories(overlapper_triggerred PUBLIC include ${CUB_DIR})
-target_link_libraries(overlapper_triggerred cgadefs cgalogging cgabase cgaio cudaaligner)
+target_link_libraries(overlapper_triggerred cgabase cgaio cgalogging cudaaligner)
 target_compile_options(overlapper_triggerred PRIVATE -Werror)
 
 add_doxygen_source_dir(${CMAKE_CURRENT_SOURCE_DIR}/include)
@@ -91,7 +91,7 @@ target_include_directories(cudamapper
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
 )
 
-target_link_libraries(cudamapper cgadefs cgabase index_gpu matcher_gpu cgalogging overlapper_triggerred cudamapper_utils)
+target_link_libraries(cudamapper cgabase cgalogging cudamapper_utils index_gpu matcher_gpu overlapper_triggerred)
 
 # Add tests folder
 add_subdirectory(tests)
